Fixed a cyclical reference
