+++ packages/utilities/adaptive-ui/src/design-tokens/type.ts
@@ -1,6 +1,22 @@
 import { DesignToken } from "@microsoft/fast-foundation";
-import { StandardFontWeight } from "../type/type-ramp.js";
 import { create } from "./create.js";
+
+/**
+ * Standard font wights.
+ *
+ * @public
+ */
+export const StandardFontWeight = {
+    Thin: 100,
+    ExtraLight: 200,
+    Light: 300,
+    Normal: 400,
+    Medium: 500,
+    SemiBold: 600,
+    Bold: 700,
+    ExtraBold: 800,
+    Black: 900,
+} as const;
 
 /** @public */
 export const bodyFont = create<string>("body-font").withDefault(
