+++ packages/utilities/adaptive-ui/src/type/type-ramp.ts
@@ -30,23 +30,6 @@ import {
     typeRampPlus6LineHeight,
 } from "../design-tokens/type.js";
 
-/**
- * Standard font wights.
- *
- * @public
- */
-export const StandardFontWeight = {
-    Thin: 100,
-    ExtraLight: 200,
-    Light: 300,
-    Normal: 400,
-    Medium: 500,
-    SemiBold: 600,
-    Bold: 700,
-    ExtraBold: 800,
-    Black: 900,
-} as const;
-
 /** @public */
 export const typeRampBase = cssPartial`
     font-family: ${bodyFont};
